examples/llm_compression/onnx/tiny_llama/requirements.txt
transformers
openvino==2025.2.0
optimum-intel[openvino]
git+https://github.com/onnx/onnx.git
onnxruntime==1.21.1
